% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compressBigwig.R
\name{compressBigwig}
\alias{compressBigwig}
\title{compressBigwig}
\usage{
compressBigwig(
  in_bigwig_path,
  out_bedgraph_path,
  partition_length = 1e+06,
  n_cores = 4,
  bin = NULL,
  ...
)
}
\arguments{
\item{in_bigwig_path}{Path to the BigWig file to be compressed.}

\item{out_bedgraph_path}{Path to the BedGraph file. The result of the partition will be saved in this file.}

\item{partition_length}{The number of loci to be read into the memory from the BigWig file. The default is 1e6.}

\item{n_cores}{The number of cores to be used in the segmentation. The default is 4.}

\item{bin}{Size of the data bins. If it has a value other than NULL, the data is binned before segmentation.}

\item{...}{To pass parameters to the L0 segmentation backend. For example, lambda2.}
}
\description{
Compresses the signal stored in the given BigWig file solving a L0 segmentation problem and stores the output into the given BedGraph file.
}
