\name{l01segmentation-package}
\alias{l01segmentation-package}
\alias{l01segmentation}
\docType{package}
\title{
  Solves L0 segmentation and L1 approximator problems for various probability distributions
}
\description{
  A more detailed description of what the package does. A length
  of about one to five lines is recommended.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Your Name, email optional.

Maintainer: Ali Tugrul Balci <alitugrulbalci@gmail.com>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
