% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotsegments}
\alias{plotsegments}
\title{plotsegments}
\usage{
plotsegments(
  segments,
  data = NULL,
  label = NULL,
  title = "",
  ylab = "",
  xlab = "",
  show = "fused"
)
}
\arguments{
\item{segments}{The segmentation outputed by fusedsegmentation (format = "compressed").}

\item{data}{The original signal vector. If NULL, only the segmentation is plotted.}

\item{label}{Labels for each location. The data points are colored by this vector.}

\item{ylab}{Y-axis label.}

\item{xlab}{X-axis label.}
}
\description{
Plots the segmentation over the data.
}
